IrDA update 2/3, set_bit updates:  <Following patch from Paul Mackerras> o [CORRECT] Argument of set_bit and friends should be unsigned long Should fix all compile warnings ;-)

+++ include/net/irda/irlmp.h
@@ -100,7 +100,7 @@ struct lsap_cb {
 	irda_queue_t queue;      /* Must be first */
 	magic_t magic;
 
-	int  connected;
+	unsigned long connected;	/* set_bit used on this */
 	int  persistent;
 
 	__u8 slsap_sel;   /* Source (this) LSAP address */
